explicit narrowing

+++ src/application/SettingsComponent.cpp
@@ -19,6 +19,7 @@
 #include <exception>
 
 #include <fmt/format.h>
+#include <gsl/gsl>
 
 #include "Misc.h"
 #include "SettingsManager.h"
@@ -121,7 +122,8 @@ void SettingsComponent::Init()
       addToLayout(&autohide_setting_, anchorMidLeft, anchorMidRight);
       addAndMakeVisible(autohide_setting_);
       autohide_setting_.onValueChange = [this] {
-         settings_manager_.SetAutoHideTime(std::lrint(autohide_setting_.getValue()));
+         settings_manager_.SetAutoHideTime(
+             gsl::narrow<int>(std::lrint(autohide_setting_.getValue())));
          rsj::Log(fmt::format(
              FMT_STRING("Autohide time set to {} seconds."), settings_manager_.GetAutoHideTime()));
       };
